pyproject.toml
[tool.poetry]
name = "trackbinance"
version = "0.1.0"
description = ""
authors = ["Abrorjon Ruziboev <abror.ruzibayev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
djangorestframework = "^3.13.1"
celery = "^5.2.6"
psycopg2-binary = "^2.9.3"
requests = "^2.27.1"
redis = "^4.3.1"
django-celery-beat = "^2.2.1"
django-celery-results = "^2.3.1"
coverage = "^6.4"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
